requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
certifi==2024.12.14
eval_type_backport==0.2.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jsonpath-python==1.0.6
magic-filter==1.0.12
mistralai==1.2.5
multidict==6.1.0
mypy-extensions==1.0.0
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
typing-inspect==0.9.0
yarl==1.18.3